added more comments
