+++ src/background/glassdoor_scraper.py
@@ -1,3 +1,27 @@
+'''
+Execution flow:
+
+Background.js
+
+Listens for: a tab change event fired when the current tabs url changes
+Executes: scrapes the jobId from the url
+Sends: a message to the contentScript that we recieved a new job
+\/
+\/
+ContentScript.js
+Listens for: the new job event from background.js
+Executes the scraping of the linkedin and glassdoor
+Calls:
+\/
+\/
+glassdoor_scraper_server
+Listens for: requests sent from content script to scrape glassdoor for a given compnay
+Executes the functions from 
+\/
+\/
+glassdoor_scraper
+'''
+
 #Credit to scrapfly
 
 """
@@ -15,17 +39,11 @@ import json
 import os
 import re
 import httpx
-from random import choice
 from bs4 import BeautifulSoup
 from html import escape
 from typing import Dict, List, Optional, Tuple, TypedDict
 from urllib.parse import urljoin
-from flask import Flask, jsonify, request
 from loguru import logger as log
-from flask_cors import CORS
-
-app = Flask(__name__)
-CORS(app)
 
 def extract_apollo_state(html):
     """Extract apollo graphql state data from HTML source"""
@@ -194,104 +212,5 @@ class Url:
             new = re.sub(".htm", f"_P{page}.htm", url)
         assert new != url
         return new
-@app.route('/get_glassdoor_data', methods=['GET'])
-async def run():
-    if request.method == 'OPTIONS':
-        print("RECIEVED OPTIONS REQUEST, PREFLIGHT")
-        response = jsonify({'status': 'ok'})
-        response.headers['Access-Control-Allow-Origin'] = '*'
-        response.headers['Access-Control-Allow-Methods'] = 'GET, POST, OPTIONS'
-        response.headers['Access-Control-Allow-Headers'] = 'Content-Type, Authorization'
-        return response
-    company = request.args.get('company', default="NO COMPANY LOADED", type=str)
-    #remove non utf8 characters
-    company = company.encode('utf-8', errors='ignore').decode('utf-8')
-    if company == "NO COMPANY LOADED":
-        raise AttributeError("Could not load company")
-    print("Company Loaded: " + company)
-    headers_list = [
-        {
-            "User-Agent": (
-                "Mozilla/5.0 (Windows NT 10.0; Win64; x64) "
-                "AppleWebKit/537.36 (KHTML, like Gecko) "
-                "Chrome/91.0.4472.124 Safari/537.36"
-            ),
-            "Accept-Language": "en-US,en;q=0.9",
-            "Accept-Encoding": "gzip, deflate, br",
-            "Connection": "keep-alive",
-            "Upgrade-Insecure-Requests": "1",
-            "Referer": "https://www.google.com/",
-            "DNT": "1"
-        },
-        {
-            "User-Agent": (
-                "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) "
-                "AppleWebKit/605.1.15 (KHTML, like Gecko) "
-                "Version/14.1.2 Safari/605.1.15"
-            ),
-            "Accept-Language": "en-US,en;q=0.8",
-            "Accept-Encoding": "gzip, deflate, br",
-            "Connection": "keep-alive",
-            "Upgrade-Insecure-Requests": "1",
-            "Referer": "https://www.apple.com/",
-            "DNT": "1"
-        },
-        {
-            "User-Agent": (
-                "Mozilla/5.0 (Windows NT 10.0; Win64; x64; rv:89.0) "
-                "Gecko/20100101 Firefox/89.0"
-            ),
-            "Accept-Language": "en-US,en;q=0.7",
-            "Accept-Encoding": "gzip, deflate, br",
-            "Connection": "keep-alive",
-            "Upgrade-Insecure-Requests": "1",
-            "Referer": "https://www.mozilla.org/",
-            "DNT": "1"
-        },
-        {
-            "User-Agent": (
-                "Mozilla/5.0 (X11; Linux x86_64) "
-                "AppleWebKit/537.36 (KHTML, like Gecko) "
-                "Chrome/90.0.4430.93 Safari/537.36"
-            ),
-            "Accept-Language": "en-US,en;q=0.9",
-            "Accept-Encoding": "gzip, deflate, br",
-            "Connection": "keep-alive",
-            "Upgrade-Insecure-Requests": "1",
-            "Referer": "https://www.google.com/",
-            "DNT": "1"
-        },
-        {
-            "User-Agent": (
-                "Mozilla/5.0 (iPhone; CPU iPhone OS 14_6 like Mac OS X) "
-                "AppleWebKit/605.1.15 (KHTML, like Gecko) "
-                "Version/14.0 Mobile/15E148 Safari/604.1"
-            ),
-            "Accept-Language": "en-US,en;q=0.9",
-            "Accept-Encoding": "gzip, deflate, br",
-            "Connection": "keep-alive",
-            "Upgrade-Insecure-Requests": "1",
-            "Referer": "https://www.apple.com/",
-            "DNT": "1"
-        }
-    ]
-    client = httpx.Client(headers=choice(headers_list), follow_redirects=True)
-    companies = await find_companies(company, client)
-    company_data_url = companies[0]["url_reviews"]
-    print("Company Data Url: "+company_data_url)
-    company_data_full = await scrape_cache(company_data_url, client)
-    print(json.dumps(company_data_full, indent=2))
-    return jsonify({
-        "overallRating": company_data_full["ratings"]["overallRating"],
-        "businessOutlookRating": company_data_full["ratings"]["businessOutlookRating"],
-        "careerOpportunitiesRating": company_data_full["ratings"]["careerOpportunitiesRating"],
-        "ceoRating": company_data_full["ratings"]["ceoRating"],
-        "compensationAndBenefitsRating": company_data_full["ratings"]["compensationAndBenefitsRating"],
-        "cultureAndValuesRating": company_data_full["ratings"]["cultureAndValuesRating"],
-        "diversityAndInclusionRating": company_data_full["ratings"]["diversityAndInclusionRating"],
-        "seniorManagementRating": company_data_full["ratings"]["seniorManagementRating"],
-        "workLifeBalanceRating": company_data_full["ratings"]["workLifeBalanceRating"]
-    })
-if __name__ == '__main__':
-    app.run(debug=True)
+
